ATOM INSPIRE: Reference database for ATOM cadastral parcels

Create a database containing the urls provided in the INSPIRE ATOM service for extracting Cadastral Parcels.

catrnav_atom_get_parcels_db_all(
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)

Source

SITNA – Catastro de Navarra

Arguments

cache

A logical whether to do caching. Default is TRUE. See About caching section on catrnav_set_cache_dir().

update_cache

A logical whether to update cache. Default is FALSE. When set to TRUE it would force a fresh download of the source file.

cache_dir

A path to a cache directory. On missing value the function would store the cached files on a temporary dir (See base::tempdir()).

verbose

Logical, displays information. Useful for debugging, default is FALSE.

Value

A tibble with the information requested:

munic: Name of the municipality.

url: url for downloading information of the corresponding municipality.

date: Reference date of the data.

See also

Examples

# \donttest{
catrnav_atom_get_parcels_db_all()
#> # A tibble: 342 × 3
#>    munic                            url                               date      
#>    <chr>                            <chr>                             <date>    
#>  1 001 Abáigar                      https://filescartografia.navarra… 2023-06-30
#>  2 002 Abárzuza / Abartzuza         https://filescartografia.navarra… 2023-06-30
#>  3 003 Abaurregaina / Abaurrea Alta https://filescartografia.navarra… 2023-06-30
#>  4 004 Abaurrepea / Abaurrea Baja   https://filescartografia.navarra… 2023-06-30
#>  5 005 Aberin                       https://filescartografia.navarra… 2023-06-30
#>  6 006 Ablitas                      https://filescartografia.navarra… 2023-06-30
#>  7 007 Adiós                        https://filescartografia.navarra… 2023-06-30
#>  8 008 Aguilar de Codés             https://filescartografia.navarra… 2023-06-30
#>  9 009 Aibar / Oibar                https://filescartografia.navarra… 2023-06-30
#> 10 010 Altsasu / Alsasua            https://filescartografia.navarra… 2023-06-30
#> # ℹ 332 more rows
# }

Site built with pkgdown 2.0.8.

Template rostemplate by dieghernan, based on Bootstrapious.